refactor: index, App, move global styles to AppContainer
