+++ package.json
@@ -36,7 +36,7 @@
     "react-modal": "^3.14.3",
     "react-redux": "^7.2.4",
     "react-router-dom": "^5.2.0",
-    "reset-jss": "^1.0.0"
+    "reset-css": "^5.0.1"
   },
   "devDependencies": {
     "@babel/core": "^7.14.6",
