+++ src/components/layout/index.js
@@ -1,6 +1,6 @@
+import AppContainer from './AppContainer/AppContainer';
 import Container from './Container/Container';
 import Header from './Header/Header';
 import ScrollToTop from './ScrollToTop/ScrollToTop';
-import Wrapper from './Wrapper/Wrapper';
 
-export { Container, Header, ScrollToTop, Wrapper };
+export { AppContainer, Container, Header, ScrollToTop };
